Split mdtoc to a reusable package
